# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation des APIs de Nekiema BTP",
        "description": "Implementation avec Swagger par Asher Services",
        "contact": {
            "email": "mianahissan@protonmail.com"
        },
        "license": {
            "name": "Nekiema Protected",
            "url": "https://www.nekiema.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Nekiema Service Server"
        }
    ],
    "paths": {
...
        "/v1/admin": {
            "post": {
                "tags": [
                    "Administrateur"
                ],
                "summary": "CREATE ADMINISTRATOR",
                "description": "CREATE ADMINISTRATOR",
                "operationId": "createAdmin",
                "requestBody": {
                    "description": "Transmettre les informations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "contact": {
                                        "type": "string"
                                    },
                                    "pseudo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération éffectuée",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise demande"
                    },
                    "404": {
                        "description": "pas trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}